requirements.txt
appdirs==1.4.3
asgiref==3.4.1
boto3==1.13.15
botocore==1.16.15
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
distlib==0.3.0
distro==1.4.0
Django==3.1.12
django-kuldeep==1.8
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.2
docutils==0.15.2
html5lib==1.0.1
idna==2.8
importlib-metadata==4.6.3
ipaddr==2.2.0
jmespath==0.10.0
jsonfield==3.1.0
kcalculator==0.0.10
lockfile==0.12.2
mysqlclient==2.0.3
packaging==20.3
pep517==0.8.2
Pillow==8.3.1
progress==1.5
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.11
webencodings==0.5.1
zipp==3.5.0
